+++ docs/examples/03_backends_ros/files/03_forward_kinematics_urdf.py
@@ -1,12 +1,12 @@
 from compas_robots import Configuration
-from compas_fab.robots.ur5 import Robot
+from compas_fab.robots import RobotLibrary
 
-robot = Robot()
+robot = RobotLibrary.ur5()
 configuration = Configuration.from_revolute_values([-2.238, -1.153, -2.174, 0.185, 0.667, 0.0])
 
 frame_WCF = robot.forward_kinematics(configuration)
 
 print("Frame in the world coordinate system")
-print(frame_WCF)
+print(str(frame_WCF.point))
 
 assert str(frame_WCF.point) == "Point(x=0.300, y=0.100, z=0.500)"