+++ docs/examples/05_backends_pybullet/files/02_inverse_kinematics.py
@@ -3,7 +3,7 @@ from compas.geometry import Frame
 from compas_fab.backends import PyBulletClient
 
 with PyBulletClient() as client:
-    urdf_filename = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
+    urdf_filename = compas_fab.get('robot_library/ur5_robot/urdf/robot_description.urdf')
     robot = client.load_robot(urdf_filename)
 
     frame_WCF = Frame([0.3, 0.1, 0.5], [1, 0, 0], [0, 1, 0])