+++ src/compas_fab/robots/__init__.py
@@ -72,6 +72,15 @@ Constraints
     OrientationConstraint
     PositionConstraint
 
+Built-in robots
+---------------
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    RobotLibrary
+
 Unit conversion
 ---------------
 
@@ -110,6 +119,9 @@ from .reachability_map import (
 )
 from .robot import (
     Robot,
+)
+from .robot_library import (
+    RobotLibrary,
 )
 from .semantics import (
     RobotSemantics,
@@ -160,6 +172,7 @@ __all__ = [
     "DeviationVectorsGenerator",
     "OrthonormalVectorsFromAxisGenerator",
     "Robot",
+    "RobotLibrary",
     "RobotSemantics",
     "Target",
     "Tool",