+++ src/compas_fab/robots/robot.py
@@ -140,8 +140,9 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.name
-        'ur5'
+        'ur5_robot'
         """
         return self.model.name
 
@@ -151,6 +152,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.group_names
         ['manipulator', 'endeffector']
 
@@ -175,6 +177,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> sorted(robot.group_states['manipulator'].keys())
         ['home', 'up']
 
@@ -206,8 +209,9 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_end_effector_link_name()
-        'ee_link'
+        'tool0'
         """
         if not self.semantics:
             return self.model.get_end_effector_link_name()
@@ -228,9 +232,10 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> link = robot.get_end_effector_link()
         >>> link.name
-        'ee_link'
+        'tool0'
         """
         name = self.get_end_effector_link_name(group)
         return self.model.get_link_by_name(name)
@@ -268,6 +273,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_base_link_name()
         'base_link'
         """
@@ -290,6 +296,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> link = robot.get_base_link()
         >>> link.name
         'base_link'
@@ -330,13 +337,14 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_link_names('manipulator')
-        ['base_link', 'shoulder_link', 'upper_arm_link', 'forearm_link', 'wrist_1_link', 'wrist_2_link', 'wrist_3_link', 'ee_link']
+        ['base_link', 'base_link_inertia', 'shoulder_link', 'upper_arm_link', 'forearm_link', 'wrist_1_link', 'wrist_2_link', 'wrist_3_link', 'flange', 'tool0']
         """
         base_link_name = self.get_base_link_name(group)
-        ee_link_name = self.get_end_effector_link_name(group)
+        end_effector_link_name = self.get_end_effector_link_name(group)
         link_names = []
-        for link in self.model.iter_link_chain(base_link_name, ee_link_name):
+        for link in self.model.iter_link_chain(base_link_name, end_effector_link_name):
             link_names.append(link.name)
         return link_names
 
@@ -349,8 +357,9 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_link_names_with_collision_geometry()
-        ['base_link', 'shoulder_link', 'upper_arm_link', 'forearm_link', 'wrist_1_link', 'wrist_2_link', 'wrist_3_link', 'ee_link']
+        ['base_link_inertia', 'shoulder_link', 'upper_arm_link', 'forearm_link', 'wrist_1_link', 'wrist_2_link', 'wrist_3_link']
         """
         return [link.name for link in self.model.iter_links() if link.collision]
 
@@ -373,6 +382,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> joints = robot.get_configurable_joints('manipulator')
         >>> [j.name for j in joints]
         ['shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint']
@@ -433,6 +443,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_configurable_joint_names('manipulator')
         ['shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', \
         'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint']
@@ -459,6 +470,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.get_configurable_joint_types('manipulator')
         [0, 0, 0, 0, 0, 0]
         """
@@ -487,6 +499,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> robot.zero_configuration('manipulator')
         Configuration((0.000, 0.000, 0.000, 0.000, 0.000, 0.000), (0, 0, 0, 0, 0, 0), \
             ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))
@@ -672,7 +685,7 @@ class Robot(Data):
         group : :obj:`str`
             The name of the planning group.
         group_state : :obj:`str`
-            The name of the `group_state`.
+            The name of the group_state.
 
         Returns
         -------
@@ -766,6 +779,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> frame_WCF = Frame([-0.363, 0.003, -0.147], [0.388, -0.351, -0.852], [0.276, 0.926, -0.256])
         >>> frame_RCF = robot.to_local_coordinates(frame_WCF)
         >>> frame_RCF                                                                                       # doctest: +SKIP
@@ -791,6 +805,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> frame_RCF = Frame([-0.363, 0.003, -0.147], [0.388, -0.351, -0.852], [0.276, 0.926, -0.256])
         >>> frame_WCF = robot.to_world_coordinates(frame_RCF)
         >>> frame_WCF                                                                                       # doctest: +SKIP
@@ -832,6 +847,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
         >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
         >>> robot.attach_tool(Tool(mesh, frame))
@@ -868,6 +884,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
         >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
         >>> robot.attach_tool(Tool(mesh, frame))
@@ -904,6 +921,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
         >>> frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
         >>> tool = Tool(mesh, frame)
@@ -1037,6 +1055,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> frame_WCF = Frame([0.3, 0.1, 0.5], [1, 0, 0], [0, 1, 0])
         >>> start_configuration = robot.zero_configuration()
         >>> group = robot.main_group_name
@@ -1112,6 +1131,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> frame_WCF = Frame([0.3, 0.1, 0.5], [1, 0, 0], [0, 1, 0])
         >>> start_configuration = robot.zero_configuration()
         >>> group = robot.main_group_name
@@ -1189,7 +1209,7 @@ class Robot(Data):
               calculate the forward kinematics for. Defaults to the group's end
               effector link.
               Backwards compatibility note: if there's no ``link`` option, the
-              planner will try also ``ee_link`` as fallback before defaulting
+              planner will try also ``tool0`` as fallback before defaulting
               to the end effector's link.
 
             There are additional options that are specific to the backend in use.
@@ -1209,6 +1229,7 @@ class Robot(Data):
 
         Examples
         --------
+        >>> robot = RobotLibrary.ur5()
         >>> configuration = Configuration.from_revolute_values([-2.238, -1.153, -2.174, 0.185, 0.667, 0.000])
         >>> group = robot.main_group_name
         >>> frame_WCF_c = robot.forward_kinematics(configuration, group)
@@ -1231,7 +1252,7 @@ class Robot(Data):
 
         # Solve with the model
         if solver == "model":
-            link = options.get("link", options.get("ee_link"))
+            link = options.get("link", options.get("tool0"))
             link = link or self.get_end_effector_link_name(group)
             if link not in self.get_link_names(group):
                 raise ValueError("Link name {} does not exist in planning group".format(link))
@@ -1296,23 +1317,22 @@ class Robot(Data):
         Examples
         --------
 
-        >>> ros = RosClient()
-        >>> ros.run()
-        >>> robot = ros.load_robot()
-        >>> frames = [Frame([0.3, 0.1, 0.5], [1, 0, 0], [0, 1, 0]),\
+        >>> with RosClient() as client:             # doctest: +SKIP
+        #: This doctest can pass locally but persistently fails on CI in GitHub. "roslibpy.core.RosTimeoutError: Failed to connect to ROS"
+        ...     robot = client.load_robot()
+        ...     frames = [Frame([0.3, 0.1, 0.5], [1, 0, 0], [0, 1, 0]),\
                       Frame([0.5, 0.1, 0.6], [1, 0, 0], [0, 1, 0])]
-        >>> start_configuration = Configuration.from_revolute_values([-0.042, 0.033, -2.174, 5.282, -1.528, 0.000])
-        >>> group = robot.main_group_name
-        >>> options = {'max_step': 0.01,\
+        ...     start_configuration = Configuration.from_revolute_values([-0.042, 0.033, -2.174, 5.282, -1.528, 0.000])
+        ...     group = robot.main_group_name
+        ...     options = {'max_step': 0.01,\
                        'jump_threshold': 1.57,\
                        'avoid_collisions': True}
-        >>> trajectory = robot.plan_cartesian_motion(frames,\
+        ...     trajectory = robot.plan_cartesian_motion(frames,\
                                                      start_configuration,\
                                                      group=group,\
                                                      options=options)
-        >>> len(trajectory.points) > 1
+        ...     len(trajectory.points) > 1
         True
-        >>> ros.close()
         """
         options = options or {}
         max_step = options.get("max_step")
@@ -1411,45 +1431,40 @@ class Robot(Data):
             The calculated trajectory.
 
         Examples
-        --------
+        --------ssssss
 
-        Using position and orientation constraints:
+        Using a :class:`~compas_fab.robots.FrameTarget` :
 
-        >>> from compas_fab.robots import FrameTarget
-        >>> ros = RosClient()
-        >>> ros.run()
-        >>> robot = ros.load_robot()
-        >>> frame = Frame([0.4, 0.3, 0.4], [0, 1, 0], [0, 0, 1])
-        >>> tolerance_position = 0.001
-        >>> tolerance_orientation = math.radians(1)
-        >>> start_configuration = Configuration.from_revolute_values([-0.042, 4.295, 0, -3.327, 4.755, 0.])
-        >>> group = robot.main_group_name
-        >>> target = FrameTarget(frame, tolerance_position, tolerance_orientation)
-        >>> trajectory = robot.plan_motion(target, start_configuration, group, {'planner_id': 'RRTConnect'})
-        >>> trajectory.fraction
+        >>> with RosClient() as client:             # doctest: +SKIP
+        #: This doctest can pass locally but persistently fails on CI in GitHub. "roslibpy.core.RosTimeoutError: Failed to connect to ROS"
+        ...     robot = client.load_robot()
+        ...     frame = Frame([0.4, 0.3, 0.4], [0, 1, 0], [0, 0, 1])
+        ...     tolerance_position = 0.001
+        ...     tolerance_orientation = math.radians(1)
+        ...     start_configuration = Configuration.from_revolute_values([-0.042, 4.295, 0, -3.327, 4.755, 0.])
+        ...     group = robot.main_group_name
+        ...     target = FrameTarget(frame, tolerance_position, tolerance_orientation)
+        ...     goal_constraints = robot.constraints_from_frame(frame, tolerance_position, tolerances_axes, group)
+        ...     trajectory = robot.plan_motion((target, start_configuration, group, {'planner_id': 'RRTConnect'})
+        ...     print(trajectory.fraction)
         1.0
-        >>> len(trajectory.points) > 1
-        True
-        >>> ros.close()
 
-        Using joint constraints (to the UP configuration):
 
-        >>> from compas_fab.robots import ConfigurationTarget
-        >>> ros = RosClient()
-        >>> ros.run()
-        >>> robot = ros.load_robot()
-        >>> joint_names = robot.get_configurable_joint_names()
-        >>> configuration = Configuration.from_revolute_values([0.0, -1.5707, 0.0, -1.5707, 0.0, 0.0], joint_names)
-        >>> tolerances_above = [math.radians(5)] * len(configuration.joint_values)
-        >>> tolerances_below = [math.radians(5)] * len(configuration.joint_values)
-        >>> group = robot.main_group_name
-        >>> target = ConfigurationTarget(configuration, tolerances_above, tolerances_below)
-        >>> trajectory = robot.plan_motion(target, start_configuration, group, {'planner_id': 'RRTConnect'})
-        >>> trajectory.fraction
+        Using a :class:`~compas_fab.robots.ConfigurationTarget` :
+
+        >>> with RosClient() as client:             # doctest: +SKIP
+        #: This doctest can pass locally but persistently fails on CI in GitHub. "roslibpy.core.RosTimeoutError: Failed to connect to ROS"
+        ...     robot = client.load_robot()
+        ...     joint_names = robot.get_configurable_joint_names()
+        ...     configuration = Configuration.from_revolute_values([0.0, -1.5707, 0.0, -1.5707, 0.0, 0.0], joint_names)
+        ...     tolerances_above = [math.radians(5)] * len(configuration.joint_values)
+        ...     tolerances_below = [math.radians(5)] * len(configuration.joint_values)
+        ...     group = robot.main_group_name
+        ...     target = ConfigurationTarget(configuration, tolerances_above, tolerances_below)
+        ...     trajectory = robot.plan_motion(target, start_configuration, group, {'planner_id': 'RRTConnect'})
+        ...     print(trajectory.fraction)
         1.0
-        >>> len(trajectory.points) > 1
-        True
-        >>> ros.close()
+
         """
         options = options or {}
         path_constraints = options.get("path_constraints")